redesign of UpcomingEventsFragment.kt

+++ app/src/main/java/ru/myitschool/nasa_bootcamp/data/dto/spaceX/capsules/Capsule.kt
@@ -9,7 +9,7 @@ class Capsule(
     @field:SerializedName("capsule_id") val capsule_id: String,
     @field:SerializedName("status") val status: String,
     @field:SerializedName("original_launch") val original_launch: String?,
-    @field:SerializedName("original_launch_unix") val original_launch_unix: Int,
+    @field:SerializedName("original_launch_unix") val original_launch_unix: Long,
     @field:SerializedName("missions") val missions: List<Mission>,
     @field:SerializedName("details") val details: String?,
     @field:SerializedName("type") val type: String,
